refactore(types): Altera o nome da pasta types para modules.
Isso aumenta a abrangência de uso dessa pasta.
